Add folders and files for excel/json scripts (#1)
The objective of these scripts is as follows:
1.) User runs the provoke.py script in the terminal and provides the
file name, sheet name, and number of headers for the excel data that will
be parsed into json. If an upload into a database (in this case Snowflake) is needed,
a target database and schema is provided.
2.) The parse.py scrip will use pandas dataframes, dictionaries, and json transformation
to transform the Excel sheet's data into a json file. If multi headers are specified, then
the logic will handle the nesting differently than a single header.
3.) The json file will be created on the local device. If the user selected to upload to a
database, the push.py script handles that by creating a series of tables and staging environments
to handle the load. Note that this is designed specifically for snowflake and would need
to be modified for other products.